feat: and/or list using '&&' and '||' (#14)
## Bug fixes
- Fixed a bug relating to eval_subshell() where some file descriptors would not close before the call to read.

## And/Or feature
Implemented the and_or command from POSIX sh. My implementation is very similar (if not identical?).

The grammar:
```
and_or          -> statement ( ( "&&" | "||" ) statement )* ;
```
It is left associative.

If the LHS statement is an expression statement then we use the truthyness of the value got from the expression statement. If the LHS is any other statement then we use the previous exit code as predicate.